Finalizamos el proyecto. Presenta errores al cargar las notas aunque reconoce las sesiones de los usuarios. Esto en parte tiene que ver con que trabajamos con VITE y el proyecto se hizo con CREATE-REACT-APP. Podriamos retomarlo luego para ver donde esta el error por el cual: No genera una cookie cuando se inicia sesion, ni tampoco carga ninguna nota a la mongodbatlas.

+++ codingInFlow-todo-app-typescript/environment/backend/src/controllers/notes.ts
@@ -2,11 +2,15 @@ import { RequestHandler } from "express";
 import noteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
-export const getNotes: RequestHandler = async (_req, res, next) => {
+export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId
+
     try {
         //throw createHttpError(401) //! Para probar el error handler
-        const notes = await noteModel.find().exec();
+        assertIsDefined(authenticatedUserId)
+        const notes = await noteModel.find({ userId: authenticatedUserId }).exec();
         res.status(200).json(notes)
     } catch (error) {
         next(error)
@@ -15,16 +19,19 @@ export const getNotes: RequestHandler = async (_req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId
-
+    const authenticatedUserId = req.session.userId
     try {
+        assertIsDefined(authenticatedUserId)
         if (!mongoose.isValidObjectId(noteId)) throw createHttpError(400, "Invalid note id")
         const note = await noteModel.findById(noteId).exec()
         if (!note) throw createHttpError(404, "Note not found.")
+        if (!note.userId.equals(authenticatedUserId)) throw createHttpError(401, "You cannot access this note.")
         res.status(200).json(note)
     } catch (error) {
         next(error)
     }
 }
+
 //* Cuando sabemos los tipos de datos qe esperamos.
 interface CreateNoteBody {
     title?: string,
@@ -34,10 +41,13 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title
     const text = req.body.text
+    const authenticatedUserId = req.session.userId
     try {
+        assertIsDefined(authenticatedUserId)
         if (!title) throw createHttpError(400, "Note must have a title.")
 
         const newNote = await noteModel.create({
+            userId: authenticatedUserId, //? El usuario qe esta logueado
             title: title,
             text: text
         })
@@ -60,13 +70,16 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId
     const newTitle = req.body.title
     const newText = req.body.text
+    const authenticatedUserId = req.session.userId
 
     try {
+        assertIsDefined(authenticatedUserId)
         if (!mongoose.isValidObjectId(noteId)) throw createHttpError(400, "Invalid note id")
         if (!newTitle) throw createHttpError(400, "Note must have a title.")
         const note = await noteModel.findById(noteId).exec()
         //* const note = await noteModel.findByIdAndUpdate(noteId).exec()  OTRA FORMA DE ACTUALIZAR
         if (!note) throw createHttpError(404, "Note not found.")
+        if (!note.userId.equals(authenticatedUserId)) throw createHttpError(401, "You cannot access this note.")
         note.title = newTitle
         note.text = newText
         const updatedNote = await note.save()
@@ -78,11 +91,13 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId
-
+    const authenticatedUserId = req.session.userId
     try {
+        assertIsDefined(authenticatedUserId)
         if (!mongoose.isValidObjectId(noteId)) throw createHttpError(400, "Invalid note id")
         const note = await noteModel.findById(noteId).exec()
         if (!note) throw createHttpError(404, "Note not found.")
+        if (!note.userId.equals(authenticatedUserId)) throw createHttpError(401, "You cannot access this note.")
         await note.deleteOne()
         res.sendStatus(204) // deleted successfully
     } catch (error) {
